docs: Use convert-kubeconfig command instead of referencing script
The functionality of the original [script](https://github.com/kubermatic/kubermatic-installer/blob/master/kubeconfig-serviceaccounts.sh) was incorporated in the `kubermatic-installer` so this change removes the external reference and provides a direct usage example of the command.

(based on the info from @xrstf, script won't be maintained and updated anymore).

Update is done in both CE/EE guides and in master + 2.17 version.