Documentation for start.kubermatic.io
Includes a proposal of the documenation for start.kubermatic.io.

Documentation is a high-level content under KKP, we are still waiting for some updated schemas.

Content is added only under `master` version (as we will iterate on top of it) and we can potentially remove that from the links in left navigation bar.

Fixes kubermatic/startio#33.