Documentation for start.kubermatic.io (#753)
* Documentation for start.kubermatic.io

Includes a proposal of the documenation for start.kubermatic.io.

Documentation is a high-level content under KKP, we are still waiting for some updated schemas.

Content is added only under `master` version (as we will iterate on top of it) and we can potentially remove that from the links in left navigation bar.

Fixes kubermatic/startio#33.

* fixup! Documentation for start.kubermatic.io

Update the regex so that spell checker is happy (anyway it should probably ignore the code blocks).

* fixup! Documentation for start.kubermatic.io

Marko's review comments - many thanks for that @xmudrii!

Updates all images to use a fixed width (as the height may be different).

* fixup! Documentation for start.kubermatic.io

* fixup! Documentation for start.kubermatic.io

* fixup! Documentation for start.kubermatic.io