Update examples for KKP v2.20 (#1025)
* Correct feature gates in examples

Adjusted all examples containing the field `featureGates`, since with
the changes to the CRDs in KKP v2.20.0 it is now of type
`object (keys:string, values:boolean)`.

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>

* Correct feature gate for user cluster MLA in KKP v2.20

Adjusted the examples in the user cluster MLA guide, since the field to
en-/disable the user cluster MLA was renamed to `userClusterMLAEnabled`
in KKP v2.20.0.

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>

* Use field requiredEmails in Seed / Preset examples

Adjusted some examples of Seeds / Presets, since the field to restrict
the usage was renamed to `requiredEmails` in KKP v2.20.0.

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>

* Correct datacenter specs in examples

Adjusted the examples of the datacenter specs to reflect the changes to
the CRDs in KKP v2.20.0.

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>

* Update examples for KKP v2.20.0

Signed-off-by: Daniel Kraus <daniel.kraus@kubermatic.com>